nest original visualize/diff under visualize/diff/stations
