+++ frontend/app/templates/visualize/diff/index.hbs
