+++ frontend/app/templates/visualize/diff/stations.hbs
@@ -1,0 +1,101 @@
+<div class="ui middle aligned stackable grid container">
+  <div class="row">
+    <div class="column">
+      <h1 class="ui huge header">
+        {{t 'visualize.diff.header'}}
+        <div class="sub header">
+          {{t 'visualize.diff.sub-header'}}
+        </div>
+      </h1>
+    </div>
+  </div>
+
+  <div class="row">
+    <div class="ui sixteen wide mobile ten wide computer column">
+      {{#ui-accordion class="fluid"}}
+        <div class="title accordion-header">
+          <i class="dropdown icon"></i>
+          {{t 'visualize.diff.explanations.header'}}
+        </div>
+        <div class="content">
+          <div id="inner-explanation-accordion" class="accordion transition hidden">
+            <div class="title accordion-header">
+              <i class="dropdown icon"></i>
+              {{t 'visualize.diff.explanations.actual-amount.header'}}
+            </div>
+            <div class="content accordion-text">
+              <p>
+              {{t 'visualize.diff.explanations.actual-amount.content'}}
+              </p>
+            </div>
+            <div class="title accordion-header">
+              <i class="dropdown icon"></i>
+              {{t 'visualize.diff.explanations.expected-amount.header'}}
+            </div>
+            <div class="content accordion-text">
+              <p>
+              {{t 'visualize.diff.explanations.expected-amount.content'}}
+              <br>
+              <div class="ui basic compact segment">
+                <strong>
+                <span>expected(broadcast) = impressions(broadcast) * average</span>
+                <br>
+                <span>average = total_amount / total_impressions </span>
+                </strong>
+              </div>
+              </p>
+            </div>
+            <div class="title accordion-header">
+              <i class="dropdown icon"></i>
+              {{t 'visualize.diff.explanations.number-of-broadcasts.header'}}
+            </div>
+            <div class="content accordion-text">
+              <p>
+              {{t 'visualize.diff.explanations.number-of-broadcasts.content'}}
+              </p>
+            </div>
+          </div>
+        </div>
+      {{/ui-accordion}}
+    </div>
+  </div>
+
+  <div class="row">
+    <div class="column">
+      <h1 class="ui large centered header">
+        {{t 'visualize.diff.charts.tv.header'}}
+      </h1>
+      <div id='chart-area'>
+        {{dynamic-high-charts chartOptions=tvChartOptions content=tvChartData}}
+      </div>
+    </div>
+  </div>
+
+  <div class="row">
+    <div class="column">
+      <div class="ui padded basic segment">
+    </div>
+    </div>
+  </div>
+
+  <div class="row">
+    <div class="column">
+      <h1 class="ui large centered header">
+        {{t 'visualize.diff.charts.radio.header'}}
+      </h1>
+      <div id='chart-area'>
+        {{dynamic-high-charts chartOptions=radioChartOptions content=radioChartData}}
+      </div>
+    </div>
+  </div>
+
+
+  <div class="row">
+    <div class="column">
+      {{#link-to 'visualize.index' class="ui huge primary button"}}
+        <i class="left arrow icon"></i>
+        {{t 'visualize.overview' }}
+      {{/link-to}}
+    </div>
+  </div>
+</div>
