+++ frontend/app/templates/visualize/index.hbs
@@ -3,7 +3,7 @@
   <div class="doubling two column row">
     <div class="column">
       <div class="ui huge header">{{t 'visualize.index.headers.diff'}}</div>
-      {{#link-to 'visualize.diff' id='visualize-diff' class="ui image"}}
+      {{#link-to 'visualize.diff.stations' id='visualize-diff' class="ui image"}}
         <img src="/assets/images/missing-features/visualize/diff.png">
       {{/link-to}}
     </div>
