+++ frontend/tests/unit/controllers/visualize/diff/stations-test.js
@@ -2,12 +2,15 @@ import { expect } from 'chai';
 import { describe, it } from 'mocha';
 import { setupTest } from 'ember-mocha';
 
-describe('Unit | Controller | visualize/diff', function() {
-  setupTest('controller:visualize/diff', {
-   needs: [
+describe('Unit | Controller | visualize/diff/stations', function() {
+  setupTest('controller:visualize/diff/stations', {
+    // Specify the other units that are required for this test.
+    needs: [
+      'service:session',
+      'service:intl',
       'service:metrics',
       'ember-metrics@metrics-adapter:piwik', // bundled adapter
-    ]
+    ]   // needs: ['controller:foo']
   });
 
   // Replace this with your real tests.
